Refactor transform selectors
